Refactor companies house validator to new namespace
