+++ lib/defra_ruby/validators.rb
@@ -1,8 +1,12 @@
 # frozen_string_literal: true
 
+require "active_model"
+
 require_relative "validators/version"
 require_relative "validators/configuration"
 require_relative "validators/companies_house_service"
+
+require_relative "validators/companies_house_number_validator"
 
 module DefraRuby
   module Validators
